+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/b2c/AADB2CConditions.java
@@ -18,6 +18,7 @@ import java.util.Map;
 /**
  * Conditions for activating AAD B2C beans.
  */
+@Deprecated
 public final class AADB2CConditions {
 
     /**