+++ sdk/spring/azure-spring-boot/src/main/java/com/azure/spring/autoconfigure/b2c/AADB2CProperties.java
@@ -23,6 +23,7 @@ import static com.azure.spring.aad.AADAuthorizationGrantType.CLIENT_CREDENTIALS;
 /**
  * Configuration properties for Azure Active Directory B2C.
  */
+@Deprecated
 @Validated
 @ConfigurationProperties(prefix = AADB2CProperties.PREFIX)
 public class AADB2CProperties implements InitializingBean {